+++ app/page.tsx
@@ -1,6 +1,5 @@
 import Footer from "../components/Footer/Footer"
 import UploadForm from "../components/Form/UploadForm/UploadForm"
-import Header from "../components/Header/Header"
 import Layout from "../components/Layout/Layout"
 import { SEO_CONFIG } from "../config"
 
@@ -25,7 +24,6 @@ export default function Home() {
         type="application/ld+json"
         dangerouslySetInnerHTML={{ __html: JSON.stringify(jsonLd) }}
       />
-      <Header />
       <UploadForm />
       <Footer />
     </Layout>